Rewrite `prop_send_recv_half_closed` to poll-api and move
